+++ src/main/java/backend/nomad/dto/member/MemberOrderResponseDto.java
@@ -25,11 +25,11 @@ public class MemberOrderResponseDto {
     private List<OrderItemResponseDto> orderItemList;
     private ReviewResponseDto reviewList;
     private OrderStatus orderStatus;
-    private Integer totalCost;
+    private Double totalCost;
     private String payMethod;
     private String orderTime;
 
-    public MemberOrderResponseDto(Long storeId, String storeName, List<OrderItemResponseDto> orderItemList, ReviewResponseDto reviewList, Integer totalCost, String payMethod, String orderTime) {
+    public MemberOrderResponseDto(Long storeId, String storeName, List<OrderItemResponseDto> orderItemList, ReviewResponseDto reviewList, Double totalCost, String payMethod, String orderTime) {
         this.storeId = storeId;
         this.storeName = storeName;
         this.orderItemList = orderItemList;
@@ -39,7 +39,7 @@ public class MemberOrderResponseDto {
         this.orderTime = orderTime;
     }
 
-    public MemberOrderResponseDto(Long memberOrderId, Long storeId, String storeName, OrderStatus orderStatus, List<OrderItemResponseDto> orderItemList, ReviewResponseDto reviewList, Integer totalCost, String payMethod, String orderTime) {
+    public MemberOrderResponseDto(Long memberOrderId, Long storeId, String storeName, OrderStatus orderStatus, List<OrderItemResponseDto> orderItemList, ReviewResponseDto reviewList, Double totalCost, String payMethod, String orderTime) {
         this.memberOrderId = memberOrderId;
         this.storeId = storeId;
         this.storeName = storeName;
